+++ src/pais/__init__.py
@@ -26,4 +26,4 @@ __all__ = [
     "Settings",
 ]
 
-__version__ = "0.8.1"
+__version__ = "0.8.2"